pyproject.toml
[project]
name = "manga-scan-grubber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.13",
    "colorama>=0.4.6",
    "flask>=3.1.0",
    "patool>=4.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
